refactor: extract XPath checks into own module
Similar to the previous refactoring, group things that change together
in the same module.
