+++ tests/test_checks_wakeup.py
@@ -13,14 +13,7 @@ from autosuspend.checks import (
     SevereCheckError,
     TemporaryCheckError,
 )
-from autosuspend.checks.wakeup import (
-    Command,
-    File,
-    Periodic,
-    SystemdTimer,
-    XPath,
-    XPathDelta,
-)
+from autosuspend.checks.wakeup import Command, File, Periodic, SystemdTimer
 
 from . import CheckTest
 from .utils import config_section
@@ -198,154 +191,3 @@ class TestSystemdTimer(CheckTest):
         }
 
         assert SystemdTimer("foo", re.compile(".*")).check(now) is now
-
-
-class TestXPath(CheckTest):
-    def create_instance(self, name: str) -> Check:
-        return XPath(name, xpath="/a", url="nourl", timeout=5)
-
-    def test_matching(self, mocker: MockFixture) -> None:
-        mock_reply = mocker.MagicMock()
-        content_property = mocker.PropertyMock()
-        type(mock_reply).content = content_property
-        content_property.return_value = '<a value="42.3"></a>'
-        mock_method = mocker.patch("requests.Session.get", return_value=mock_reply)
-
-        url = "nourl"
-        assert XPath("foo", xpath="/a/@value", url=url, timeout=5).check(
-            datetime.now(timezone.utc)
-        ) == datetime.fromtimestamp(42.3, timezone.utc)
-
-        mock_method.assert_called_once_with(url, timeout=5, headers=None)
-        content_property.assert_called_once_with()
-
-    def test_not_matching(self, mocker: MockFixture) -> None:
-        mock_reply = mocker.MagicMock()
-        content_property = mocker.PropertyMock()
-        type(mock_reply).content = content_property
-        content_property.return_value = "<a></a>"
-        mocker.patch("requests.Session.get", return_value=mock_reply)
-
-        assert (
-            XPath("foo", xpath="/b", url="nourl", timeout=5).check(
-                datetime.now(timezone.utc)
-            )
-            is None
-        )
-
-    def test_not_a_string(self, mocker: MockFixture) -> None:
-        mock_reply = mocker.MagicMock()
-        content_property = mocker.PropertyMock()
-        type(mock_reply).content = content_property
-        content_property.return_value = "<a></a>"
-        mocker.patch("requests.Session.get", return_value=mock_reply)
-
-        with pytest.raises(TemporaryCheckError):
-            XPath("foo", xpath="/a", url="nourl", timeout=5).check(
-                datetime.now(timezone.utc)
-            )
-
-    def test_not_a_number(self, mocker: MockFixture) -> None:
-        mock_reply = mocker.MagicMock()
-        content_property = mocker.PropertyMock()
-        type(mock_reply).content = content_property
-        content_property.return_value = '<a value="narf"></a>'
-        mocker.patch("requests.Session.get", return_value=mock_reply)
-
-        with pytest.raises(TemporaryCheckError):
-            XPath("foo", xpath="/a/@value", url="nourl", timeout=5).check(
-                datetime.now(timezone.utc)
-            )
-
-    def test_multiple_min(self, mocker: MockFixture) -> None:
-        mock_reply = mocker.MagicMock()
-        content_property = mocker.PropertyMock()
-        type(mock_reply).content = content_property
-        content_property.return_value = """
-            <root>
-                <a value="40"></a>
-                <a value="10"></a>
-                <a value="20"></a>
-            </root>
-        """
-        mocker.patch("requests.Session.get", return_value=mock_reply)
-
-        assert XPath("foo", xpath="//a/@value", url="nourl", timeout=5).check(
-            datetime.now(timezone.utc)
-        ) == datetime.fromtimestamp(10, timezone.utc)
-
-    def test_create(self) -> None:
-        check: XPath = XPath.create(
-            "name",
-            config_section(
-                {
-                    "xpath": "/valid",
-                    "url": "nourl",
-                    "timeout": "20",
-                }
-            ),
-        )  # type: ignore
-        assert check._xpath == "/valid"
-
-
-class TestXPathDelta(CheckTest):
-    def create_instance(self, name: str) -> Check:
-        return XPathDelta(name, xpath="/a", url="nourl", timeout=5, unit="days")
-
-    @pytest.mark.parametrize(
-        ("unit", "factor"),
-        [
-            ("microseconds", 0.000001),
-            ("milliseconds", 0.001),
-            ("seconds", 1),
-            ("minutes", 60),
-            ("hours", 60 * 60),
-            ("days", 60 * 60 * 24),
-            ("weeks", 60 * 60 * 24 * 7),
-        ],
-    )
-    def test_smoke(self, mocker: MockFixture, unit: str, factor: float) -> None:
-        mock_reply = mocker.MagicMock()
-        content_property = mocker.PropertyMock()
-        type(mock_reply).content = content_property
-        content_property.return_value = '<a value="42"></a>'
-        mocker.patch("requests.Session.get", return_value=mock_reply)
-
-        url = "nourl"
-        now = datetime.now(timezone.utc)
-        result = XPathDelta(
-            "foo", xpath="/a/@value", url=url, timeout=5, unit=unit
-        ).check(now)
-        assert result == now + timedelta(seconds=42) * factor
-
-    def test_create(self) -> None:
-        check = XPathDelta.create(
-            "name",
-            config_section(
-                {
-                    "xpath": "/valid",
-                    "url": "nourl",
-                    "timeout": "20",
-                    "unit": "weeks",
-                }
-            ),
-        )
-        assert check._unit == "weeks"
-
-    def test_create_wrong_unit(self) -> None:
-        with pytest.raises(ConfigurationError):
-            XPathDelta.create(
-                "name",
-                config_section(
-                    {
-                        "xpath": "/valid",
-                        "url": "nourl",
-                        "timeout": "20",
-                        "unit": "unknown",
-                    }
-                ),
-            )
-
-    def test_init_wrong_unit(self) -> None:
-        with pytest.raises(ValueError, match=r".*unit.*"):
-            XPathDelta("name", url="url", xpath="/a", timeout=5, unit="unknownunit")
